extract json-utils.go
